3_TCGbda/requirements.txt
google-cloud-bigquery>=3.11.0
pandas>=1.5.0
python-dotenv>=1.0.0
pyarrow>=10.0.0